--- Frontend/src/app/auth/auth.guard.ts
@@ -1,13 +1,0 @@
-import { inject } from '@angular/core';
-import { AuthService } from './auth.service';
-import { Router, CanMatchFn } from '@angular/router';
-import { paths } from '../app.routes';
-
-export const authGuard: CanMatchFn = () => {
-  const authService = inject(AuthService);
-  const router = inject(Router);
-
-  if (authService.isLoggedIn) return true;
-
-  return router.parseUrl(paths.LOGIN);
-};
