+++ Frontend/src/app/core/components/navbar/navbar.component.html
@@ -1,0 +1,17 @@
+<mat-toolbar class="justify-end sticky top-0 z-[101]">
+    <section class="flex items-center space-x-1">
+        <button mat-icon-button>
+            <mat-icon>notifications</mat-icon>
+        </button>
+        <button mat-icon-button [matMenuTriggerFor]="subMenu">
+            <mat-icon>account_circle</mat-icon>
+        </button>
+    </section>
+</mat-toolbar>
+
+<mat-menu #subMenu="matMenu">
+    <button mat-menu-item (click)="logout()">
+        <mat-icon>logout</mat-icon>
+        <span>Cerrar sesión</span>
+    </button>
+</mat-menu>
