+++ Frontend/src/app/features/formats/components/format-row/format-row.component.ts
