+++ Frontend/src/app/features/home/home.component.scss
@@ -1,0 +1,8 @@
+.sidenav-container {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    background: white;
+}
