+++ Frontend/src/app/features/home/home.component.ts
@@ -1,0 +1,24 @@
+import { Component } from '@angular/core';
+import { RouterLink, RouterOutlet } from '@angular/router';
+import { MaterialModule } from '@shared/modules/material/material.module';
+import { FormatsComponent } from '../formats/formats.component';
+import { NavbarComponent } from '@core/components/navbar/navbar.component';
+import { paths } from 'src/app/app.routes';
+
+@Component({
+  selector: 'app-home',
+  standalone: true,
+  imports: [
+    RouterOutlet, RouterLink, MaterialModule,
+    NavbarComponent, FormatsComponent,
+  ],
+  templateUrl: './home.component.html',
+  styleUrl: './home.component.scss',
+})
+export default class HomeComponent {
+  readonly administration = [
+    { id: 'roles', path: paths.ROLES, title: 'Roles', icon: 'admin_panel_settings' },
+    { id: 'users', path: paths.USERS, title: 'Usuarios', icon: 'group' },
+    { id: 'parameters', path: paths.PARAMETERS, title: 'Parametros', icon: 'fact_check' },
+  ];
+}
