+++ Frontend/src/app/features/login/login.component.html
@@ -1,0 +1,33 @@
+<div class="flex flex-col justify-center items-center min-h-screen">
+  <a mat-button routerLink="/" class="mb-5">
+    <h3 class="mb-0">Habilident</h3>
+  </a>
+  <mat-card class="max-w-md p-8">
+    <mat-card-header class="justify-center mb-2">
+      <mat-card-title>
+        <h5>Inicie sesión en su cuenta</h5>
+      </mat-card-title>
+    </mat-card-header>
+    <mat-card-content>
+      <form [formGroup]="loginForm" (ngSubmit)="login()" class="space-y-6">
+        <mat-form-field appearance="outline" class="w-full">
+          <mat-label>Correo electrónico</mat-label>
+          <input matInput type="email" id="email" formControlName="email" required>
+          @if (loginForm.controls.email.invalid) {
+          <mat-error>Introduce un correo válido</mat-error>
+          }
+        </mat-form-field>
+
+        <mat-form-field appearance="outline" class="w-full">
+          <mat-label>Contraseña</mat-label>
+          <input matInput type="password" id="password" formControlName="password" required>
+          @if (loginForm.controls.password.invalid) {
+          <mat-error>Introduce una contraseña válida</mat-error>
+          }
+        </mat-form-field>
+
+        <button mat-flat-button type="submit" class="w-full">Iniciar sesión</button>
+      </form>
+    </mat-card-content>
+  </mat-card>
+</div>
