+++ Frontend/src/app/features/records/components/record-table/record-table.component.ts
