+++ Frontend/src/app/features/records/components/record-text/record-text.component.ts
