+++ Frontend/src/app/features/records/components/record/record.component.html
