+++ Frontend/src/app/features/records/records.component.html
