--- Frontend/src/app/formats/formats.component.scss
@@ -1,8 +1,0 @@
-.sidenav-container {
-    position: absolute;
-    top: 64px;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    background: #eee;
-}
