--- Frontend/src/app/home/home.component.html
@@ -1,37 +1,0 @@
-<mat-toolbar class="justify-between">
-  <h3 class="text-sky-600 mb-0">Habilident</h3>
-
-  <section>
-    @for (item of menu; track item.id) {
-    <button mat-button [routerLink]="item.path" routerLinkActive="active">{{item.title}}</button>
-    }
-  </section>
-
-  <section class="flex items-center space-x-1">
-    <button mat-icon-button>
-      <mat-icon>notifications</mat-icon>
-    </button>
-    <span>{{username | titlecase}}</span>
-    <button mat-icon-button [matMenuTriggerFor]="subMenu">
-      <mat-icon>more_vert</mat-icon>
-    </button>
-  </section>
-</mat-toolbar>
-
-<mat-menu #subMenu="matMenu">
-  <button mat-menu-item>
-    <mat-icon>account_circle</mat-icon>
-    <span>Perfil</span>
-  </button>
-  @for (item of administration; track item.id) {
-  <button mat-menu-item [routerLink]="item.path" routerLinkActive="active">
-    <mat-icon>{{item.icon}}</mat-icon>
-    <span>{{item.title}}</span>
-  </button>
-  }
-  <button mat-menu-item (click)="logout()">
-    <mat-icon>logout</mat-icon>
-    <span>Cerrar sesión</span>
-  </button>
-</mat-menu>
-<router-outlet />
