+++ Frontend/tsconfig.json
@@ -28,9 +28,15 @@
       "@tipos/*": [
         "../Backend/src/types/*"
       ],
+      "@core/*": [
+        "src/app/core/*"
+      ],
+      "@features/*": [
+        "src/app/features/*"
+      ],
       "@shared/*": [
         "src/app/shared/*"
-      ]
+      ],
     },
   },
   "angularCompilerOptions": {
